requirements.txt
asgiref==3.2.7
Django==3.0.7
PyMySQL==0.9.3
pytz==2020.1
sqlparse==0.3.1
djangorestframework==3.11.0
requests==2.24.0